fix: wallet connection analytics
