+++ src/components/WalletModal/index.tsx
@@ -1,6 +1,5 @@
-import { sendAnalyticsEvent, user } from '@uniswap/analytics'
-import { CustomUserProperties, InterfaceEventName, WalletConnectionResult } from '@uniswap/analytics-events'
-import { getWalletMeta } from '@uniswap/conedison/provider/meta'
+import { sendAnalyticsEvent } from '@uniswap/analytics'
+import { InterfaceEventName, WalletConnectionResult } from '@uniswap/analytics-events'
 import { useWeb3React } from '@web3-react/core'
 import { sendEvent } from 'components/analytics'
 import { AutoColumn } from 'components/Column'
@@ -8,7 +7,6 @@ import { AutoRow } from 'components/Row'
 import { useWalletDrawer } from 'components/WalletDropdown'
 import IconButton from 'components/WalletDropdown/IconButton'
 import { Connection, ConnectionType, getConnections, networkConnection } from 'connection'
-import { useGetConnection } from 'connection'
 import { ErrorCode } from 'connection/utils'
 import { isSupportedChain } from 'constants/chains'
 import { useMgtmEnabled } from 'featureFlags/flags/mgtm'
@@ -16,7 +14,6 @@ import { useCallback, useEffect, useRef, useState } from 'react'
 import { Settings } from 'react-feather'
 import { useAppDispatch } from 'state/hooks'
 import { updateSelectedWallet } from 'state/user/reducer'
-import { useConnectedWallets } from 'state/wallets/hooks'
 import styled from 'styled-components/macro'
 import { ThemedText } from 'theme'
 import { flexColumnNoWrap } from 'theme/styles'
@@ -47,32 +44,6 @@ const PrivacyPolicyWrapper = styled.div`
   padding: 0 4px;
 `
 
-const sendAnalyticsEventAndUserInfo = (
-  account: string,
-  walletType: string,
-  chainId: number | undefined,
-  isReconnect: boolean,
-  peerWalletAgent: string | undefined
-) => {
-  // User properties *must* be set before sending corresponding event properties,
-  // so that the event contains the correct and up-to-date user properties.
-  user.set(CustomUserProperties.WALLET_ADDRESS, account)
-  user.set(CustomUserProperties.WALLET_TYPE, walletType)
-  user.set(CustomUserProperties.PEER_WALLET_AGENT, peerWalletAgent ?? '')
-  if (chainId) {
-    user.postInsert(CustomUserProperties.ALL_WALLET_CHAIN_IDS, chainId)
-  }
-  user.postInsert(CustomUserProperties.ALL_WALLET_ADDRESSES_CONNECTED, account)
-
-  sendAnalyticsEvent(InterfaceEventName.WALLET_CONNECT_TXN_COMPLETED, {
-    result: WalletConnectionResult.SUCCEEDED,
-    wallet_address: account,
-    wallet_type: walletType,
-    is_reconnect: isReconnect,
-    peer_wallet_agent: peerWalletAgent,
-  })
-}
-
 function didUserReject(connection: Connection, error: any): boolean {
   return (
     error?.code === ErrorCode.USER_REJECTED_REQUEST ||
@@ -83,16 +54,13 @@ function didUserReject(connection: Connection, error: any): boolean {
 
 export default function WalletModal({ openSettings }: { openSettings: () => void }) {
   const dispatch = useAppDispatch()
-  const { connector, account, chainId, provider } = useWeb3React()
+  const { connector, chainId } = useWeb3React()
   const [drawerOpen, toggleWalletDrawer] = useWalletDrawer()
 
-  const [connectedWallets, addWalletToConnectedWallets] = useConnectedWallets()
-  const [lastActiveWalletAddress, setLastActiveWalletAddress] = useState<string | undefined>(account)
   const [pendingConnection, setPendingConnection] = useState<Connection | undefined>()
   const [pendingError, setPendingError] = useState<any>()
 
   const connections = getConnections()
-  const getConnection = useGetConnection()
 
   useEffect(() => {
     // Clean up errors when the dropdown closes
@@ -112,28 +80,6 @@ export default function WalletModal({ openSettings }: { openSettings: () => void
       networkConnection.connector.activate(chainId)
     }
   }, [chainId, connector])
-
-  // When new wallet is successfully set by the user, trigger logging of Amplitude analytics event.
-  useEffect(() => {
-    if (account && account !== lastActiveWalletAddress) {
-      const walletName = getConnection(connector).getName()
-      const peerWalletAgent = provider ? getWalletMeta(provider)?.agent : undefined
-      const isReconnect =
-        connectedWallets.filter((wallet) => wallet.account === account && wallet.walletType === walletName).length > 0
-      sendAnalyticsEventAndUserInfo(account, walletName, chainId, isReconnect, peerWalletAgent)
-      if (!isReconnect) addWalletToConnectedWallets({ account, walletType: walletName })
-    }
-    setLastActiveWalletAddress(account)
-  }, [
-    connectedWallets,
-    addWalletToConnectedWallets,
-    lastActiveWalletAddress,
-    account,
-    connector,
-    chainId,
-    provider,
-    getConnection,
-  ])
 
   // Used to track the state of the drawer in async function
   const drawerOpenRef = useRef(drawerOpen)
